refactor(FLA): colocate benchmark reference with chunk operator
